+++ NEWS.md
@@ -1,5 +1,8 @@
 # bayesplot (development version)
 
+* Add option to display observed data as points in `ppc_km_overlay()` and
+  `ppc_km_overlay_grouped()` by @Sakuski (#560)
+
 # bayesplot 1.16.0
 
 ### New plots and plotting capabilities